+++ packages/turbo-mount/src/plugins/svelte/index.ts
@@ -1,0 +1,10 @@
+import {Plugin} from "turbo-mount";
+
+import {TurboMountSvelteController} from "./turbo-mount-svelte-controller";
+
+const plugin: Plugin = {
+    framework: "svelte",
+    controller: TurboMountSvelteController
+}
+
+export default plugin;