+++ packages/turbo-mount/src/plugins/svelte/turbo-mount-svelte-controller.ts
@@ -1,13 +1,12 @@
-import {TurboMountController} from "./turbo-mount-controller";
-
 import {ComponentType} from "svelte";
+import {TurboMountController} from "turbo-mount";
 
 export class TurboMountSvelteController extends TurboMountController<ComponentType> {
     framework = "svelte"
 
-    async mountComponent(el: Element, Component: ComponentType, props: object) {
+    mountComponent(el: Element, Component: ComponentType, props: object) {
         const component = new Component({ target: el, props })
 
-        return component.$destroy
+        return () => { component.$destroy() }
     }
 }