+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
 
+### Added
+
+- New API with plugins. ([@skryukov])
+
 ## [0.1.0] - 2024-05-07
 
 ### Added